comming toplist association metabox change

+++ toplist-association-metabox.php
@@ -61,6 +61,7 @@ class TopListAssociationMetaBox extends CodeBox\CodeBox
 
     // add more if needed
     $toplists = $_POST['toplist_item_toplist'];
+    if (empty($toplist)) return;
 
     delete_post_meta($post_id, 'toplist_item_toplist');
